+++ frontend/app/components/shared/LiveSessionList/LiveSessionList.tsx
@@ -47,26 +47,18 @@ function LiveSessionList() {
 
     const _filter = { ...filter };
     let shouldUpdate = false;
-
-    // Set default sort if not already set
     if (sortOptions[1] && !filter.sort) {
       _filter.sort = sortOptions[1].value;
       shouldUpdate = true;
     }
-
-    // Only update filters if there's a change
     if (shouldUpdate) {
       searchStoreLive.edit(_filter);
     }
-
-    // Start auto-refresh timeout
     timeout();
-
-    // Cleanup on component unmount or re-run
     return () => {
       clearTimeout(timeoutId);
     };
-  }, [metaListLoading, filter.sort]); // Add necessary dependencies
+  }, [metaListLoading, filter.sort]);
 
   const refetch = () => {
     void searchStoreLive.fetchSessions();
@@ -172,7 +164,6 @@ function LiveSessionList() {
                 </Button>
               </div>
             }
-            // image={<img src="/assets/img/live-sessions.png" style={{ width: '70%', marginBottom: '30px' }} />}
             show={!loading && list.length === 0}
           >
             <div>